+++ lib/Data/qaryData.dart
@@ -11,13 +11,13 @@ String qaryDataToJson(List<QaryData> data) => json.encode(List<dynamic>.from(dat
 class QaryData {
   final String qaryName;
   final int qaryAge;
-  final int degree;
+  final double degree;
   final String testName;
   final int questions;
   QaryData(this.qaryName, this.qaryAge, this.degree, this.testName, this.questions);
 
   factory QaryData.fromJson(Map<String, dynamic> json) => QaryData(json['qaryName'],json['qaryAge'],json['degree'],json['testName'],json['questions']);
-  factory QaryData.fromFields(String name,int age,int deg,String tname,int quest)=>QaryData(name, age,deg,tname,quest);
+  factory QaryData.fromFields(String name,int age,double deg,String tname,int quest)=>QaryData(name, age,deg,tname,quest);
 
   Map<String, dynamic> toJson() => {
   };
